Changed button interrupt to be rtc interrupt
