+++ src/Components/Power.hpp
@@ -42,18 +42,27 @@
 
 extern volatile unsigned long rtcInterruptStart;
 extern volatile bool alarmTriggered;
+extern volatile unsigned long buttonInterruptStart;
+extern volatile bool buttonTriggered;
+extern volatile int buttonFlag;
+bool buttonCanTrigger;
 extern void rtc_isr();
 
 class Power : public KPComponent {
 public:
     DS3232RTC rtc;
     std::function<void()> interruptCallback;
+	std::function<void()> buttonCallback;
 
     Power(const char * name) : KPComponent(name), rtc(false) {}
 
     void onInterrupt(std::function<void()> callbcak) {
         interruptCallback = callbcak;
     }
+	
+	void onButtonInterrupt(std::function<void()> callback){
+		buttonCallback = callback;
+	}
 
     void setupRTC() {
         // Initilize RTC I2C Bus
@@ -72,6 +81,7 @@ public:
         print("RTC startup time: ");
         printCurrentTime();
         println();
+		
     }
 
     void setup() override {
@@ -82,6 +92,10 @@ public:
         // so we need this internal one. INPUT_PULLUP is required.
         pinMode(HardwarePins::RTC_INTERRUPT, INPUT_PULLUP);
         pinMode(HardwarePins::POWER_MODULE, OUTPUT);
+		
+		//init button pin
+		pinMode(HardwarePins::BUTTON_PIN, INPUT);
+		buttonCanTrigger = true;
     }
 
     /** ────────────────────────────────────────────────────────────────────────────
@@ -89,10 +103,15 @@ public:
 
      *  ──────────────────────────────────────────────────────────────────────────── */
     void update() override {
-        if (!alarmTriggered || !interruptCallback) {
+        if (!alarmTriggered || !buttonTriggered || !interruptCallback) {
             return;
         }
 
+		if(buttonFlag != 0 && buttonCanTrigger) {
+            buttonCanTrigger = false;
+            buttonTriggered = false;
+            buttonCallback();
+			}
         // Check if the interrupt is comming from RTC
         // This is important in noisy environment
         if (rtc.alarm(1) || rtc.alarm(2)) {
@@ -293,4 +312,9 @@ public:
         constexpr time_t FUDGE = 10;
         return t + FUDGE;
     }
+	
+	void setSampleButton() {
+		buttonFlag = 0;
+		buttonCanTrigger = true;
+	}
 };
